+++ src/configuration/config.py
@@ -1,5 +1,4 @@
 # ALG
-DEFAULT = True
 DEFAULT_TASK_VALUES = [
         [2, 4, 7],
         [1, 1, 2],
@@ -34,3 +33,4 @@ COLOR_PALETTE = [
         "#00f5d4", #niebieski
         "#25a18e" #zielony
 ]
+EMPTY_COLOR = "grey"