+++ src/one_app_to_rule_them_all.py
@@ -2,9 +2,8 @@ import sys
 import random
 from functools import partial
 
-from src.configuration.config import DEFAULT, DEFAULT_TASK_VALUES, T_MAX, T_MIN, N
-
-from src.configuration.config import WINDOW_HEIGHT, WINDOW_WIDTH, COLOR_PALETTE, BUTTON_SIZE, LINE_EDIT_SIZE, TASK_LABEL_WIDTH, APP_BG_COLOR, BUTTON_COLOR
+from src.configuration.config import DEFAULT_TASK_VALUES
+from src.configuration.config import WINDOW_HEIGHT, WINDOW_WIDTH, COLOR_PALETTE, EMPTY_COLOR, BUTTON_SIZE, LINE_EDIT_SIZE, TASK_LABEL_WIDTH, APP_BG_COLOR, BUTTON_COLOR
 from PyQt6.QtCore import Qt
 from PyQt6.QtWidgets import (
     QApplication,
@@ -14,7 +13,6 @@ from PyQt6.QtWidgets import (
     QPushButton,
     QLabel,
     QVBoxLayout,
-    QHBoxLayout,
     QGridLayout,
     QFormLayout,
 )
@@ -116,10 +114,6 @@ class LiuWindow(QMainWindow):
         self.user_quarter_layout.addWidget(self.button_run_alg, alignment=Qt.AlignmentFlag.AlignCenter)
 
         # add user's layout to user widget
-        # add user widget to general layout
-        # ####user_widget = QWidget()
-        # ####user_widget.setLayout(self.user_quarter_layout)
-        # ####self.general_layout.addWidget(user_widget, 0, 0)
         self.general_layout.addLayout(self.user_quarter_layout, 0, 0, alignment=Qt.AlignmentFlag.AlignCenter)
 
     def _set_input_quarter(self):
@@ -134,13 +128,7 @@ class LiuWindow(QMainWindow):
         self.input_quarter_layout.addWidget(QLabel("<h2>Input values</h2>"), alignment=Qt.AlignmentFlag.AlignCenter)
         self.input_quarter_layout.addWidget(input_values_widget, alignment=Qt.AlignmentFlag.AlignCenter)
 
-        # self.user_quarter_layout.addLayout(self.input_values_layout)
-
         # add input layout to input widget
-        # add input widget to general layout
-        # #### input_widget = QWidget()
-        # #### input_widget.setLayout(self.input_quarter_layout)
-        # #### self.general_layout.addWidget(input_widget, 1, 0)
         self.general_layout.addLayout(self.input_quarter_layout, 1, 0)
 
     def _set_output_quarter(self):
@@ -187,7 +175,7 @@ class LiuWindow(QMainWindow):
         tasks_list = []
         for el in range(tasks_number):
             execution_time = random.randint(min_time, max_time)
-            release_time = random.randint(min_time, max_time)
+            release_time = random.randint(min_time, 2*max_time)
             deadline = release_time + execution_time + random.randint(min_time, max_time)
 
             tasks_list.append(Task(p=execution_time, r=release_time, d=deadline))
@@ -277,7 +265,8 @@ class LiuWindow(QMainWindow):
                 print(f"running: {chosen_task}")
                 tasks_exec_order.append(chosen_task)
                 # set "start" time for chosen task (if not set only)
-                if not chosen_task.start_time:
+                if chosen_task.start_time is None:
+                    print(f"task {chosen_task}, setting start time to: {T}")
                     chosen_task.start_time = T
                 # run task for 1s
                 chosen_task.current_execution_time += 1
@@ -286,6 +275,8 @@ class LiuWindow(QMainWindow):
                 if chosen_task.current_execution_time == chosen_task.execution_time:
                     chosen_task.not_done = False
                     chosen_task.stop_time = T + 1  # +1 because task was executed for 1 time unit
+            else:
+                tasks_exec_order.append("---")
 
             # end of the loop
             T += 1
@@ -317,20 +308,32 @@ class LiuWindow(QMainWindow):
         print("[End]   fill_output_grid")
 
     def fill_graph_layout(self, tasks_exec_order, lmax):
+        print("[Start] fill_graph_layout")
         for idx, task in enumerate(tasks_exec_order):
+            # add task to tasks row
             graph_cell = QLabel(f"<h2>{str(task)}</h2>")
             graph_cell.setAlignment(Qt.AlignmentFlag.AlignCenter)
             # graph_cell.setFixedWidth(TASK_LABEL_WIDTH)
-            graph_cell.setStyleSheet(f"background-color: {task.task_color}")
+            if isinstance(task, Task):
+                graph_cell.setStyleSheet(f"background-color: {task.task_color}")
+            if isinstance(task, str):
+                graph_cell.setStyleSheet(f"background-color: {EMPTY_COLOR}")
             self.graph_values_layout.addWidget(graph_cell, 0, idx)
+
+            # add time index
+            time_label = QLabel(f"<p>{idx}</p>")
+            time_label.setAlignment(Qt.AlignmentFlag.AlignLeft)
+            self.graph_values_layout.addWidget(time_label, 1, idx)
+
         self.graph_quarter_layout.addWidget(QLabel(f"<h2>L<sub>max</sub>: {lmax}</h2>"), alignment=Qt.AlignmentFlag.AlignCenter)
+        print("[End]   fill_graph_layout")
 
     def order_tasks(self):
         # run Liu algorithm
         tasks_exec_order = self.perform_alg(self.tasks_list)
         # get tasks Li values
         for task in self.tasks_list:
-            task.delay = task.stop_time - task.deadline
+            task.delay = task.stop_time - task.deadline if task.stop_time > task.deadline else 0  # delay can't be negative
         # get Lmax
         lmax = max(task.delay for task in self.tasks_list)
         # prepare data for printing
@@ -340,7 +343,7 @@ class LiuWindow(QMainWindow):
             tasks_print_table.append(task_row)
         # print data
         self.fill_output_grid(tasks_print_table)
-        print(tasks_exec_order)
+        print(f"Tasks order: {tasks_exec_order}")
         self.fill_graph_layout(tasks_exec_order, lmax)
 
     def _connect_signals_and_slots(self):